+++ examples/hands_detection.py
@@ -7,9 +7,10 @@ def main():
     Vilib.display(local=True,web=True)
     Vilib.hands_detect_switch(True)
 
-    while True:
-        print(Vilib.detect_obj_parameter['hands_joints'])
-        sleep(1)
+    # while True:
+    #     print(Vilib.detect_obj_parameter['hands_joints']) # Print finger joint coordinates
+    #     sleep(0.5)
 
 if __name__ == "__main__":
     main()
+