+++ examples/pose_detection.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-import os
 from vilib import Vilib
 
 
@@ -11,3 +10,5 @@ def main():
 
 if __name__ == "__main__":
     main()
+
+