+++ examples/record_video.py
@@ -2,12 +2,17 @@
 from time import sleep,strftime,localtime
 from vilib import Vilib
 import readchar 
+from pathlib import Path
+
+import os
+user = os.getlogin()
+user_home = os.path.expanduser(f'~{user}')
 
 manual = '''
 Press keys on keyboard to control recording:
     Q: record/pause/continue
     E: stop
-    ESC: Quit
+    Ctrl^C: Quit
 '''
 
 def print_overwrite(msg,  end='', flush=True):
@@ -17,7 +22,7 @@ def print_overwrite(msg,  end='', flush=True):
 def main():
     rec_flag = 'stop' # start,pause,stop
     vname = None
-    Vilib.rec_video_set["path"] = "/home/pi/Videos/" # set path
+    Vilib.rec_video_set["path"] = f"{user_home}/Videos/" # set path
 
     Vilib.camera_start(vflip=False,hflip=False) 
     Vilib.display(local=True,web=True)
@@ -26,8 +31,7 @@ def main():
     print(manual)
     while True:
         # read keyboard
-        key = readchar.readkey()
-        key = key.lower()
+        key = readchar.readkey().lower()
         # start,pause
         if key == 'q':
             key = None
@@ -38,6 +42,7 @@ def main():
                 Vilib.rec_video_set["name"] = vname
                 # start record
                 Vilib.rec_video_run()
+                Vilib.rec_video_start()
                 print_overwrite('rec start ...')
             elif rec_flag == 'start':
                 rec_flag = 'pause'
@@ -54,7 +59,7 @@ def main():
             Vilib.rec_video_stop()
             print_overwrite("The video saved as %s%s.avi"%(Vilib.rec_video_set["path"],vname),end='\n')  
         # quit
-        elif key == readchar.key.CTRL_C or key in readchar.key.ESCAPE_SEQUENCES:
+        elif key == readchar.key.CTRL_C:
             Vilib.camera_close()
             print('\nquit')
             break 
@@ -63,3 +68,4 @@ def main():
 
 if __name__ == "__main__":
     main()
+