+++ examples/take_photo.py
@@ -1,21 +1,26 @@
 #!/usr/bin/env python3
 import time
 from vilib import Vilib
+import os
+user = os.getlogin()
+user_home = os.path.expanduser(f'~{user}')
 
 def main():
     Vilib.camera_start(vflip=False,hflip=False)
-    Vilib.display()
+    Vilib.display(local=True,web=True)
 
-    i = 0
-    path = "/home/pi/picture/vilib/photos"
+    path =f"{user_home}/Pictures/vilib/photos"
   
     while True:
         if input() == 'q': 
             _time = time.strftime("%y-%m-%d_%H-%M-%S", time.localtime())
             Vilib.take_photo(str(_time),path)
-            print("take_photo:%s.jpg"%_time)
+            print("The photo save as:%s/%s.jpg"%(path, _time))
             time.sleep(0.1)
 
 
 if __name__ == "__main__":
     main()
+
+
+