+++ vilib/__init__.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
 from .vilib import Vilib
+from .version import __version__