+++ CHANGELOG.md
@@ -1,6 +1,39 @@
 # Change Log 
 
-## [Unreleased]
+## [0.0.6] - 2023-8-16
+
+### Fixed
+- Fix problem of installation failure, and optimize the install script.
+
+### Added
+- Add checking platform when import mediapipe
+
+
+## [0.0.4] - 2022-5-19
+
+### Fixed
+- Increase compatibility for pi3 and pi4
+- Fix bug
+- Fixed use of setuptools and opencv-contrib-python version to avoid problems caused by new version
+
+
+## [0.0.3] - 2022-5-19
+
+### Fixed
+- Compatible with python version 3.9
+- Fix display bug
+- Fix detect_obj_parameter error when close
+- Fix the bug that only the first video is valid
+
+### Added
+- Add install working tip
+- Uniform version file
+
+### Optimized
+- Optimize some details of examples
+
+
+## [0.0.2] - 2022-1-14
 
 ### Added
 - Add change log